PoojaBooking service
Updated pooja Booking services and also updated the code

+++ bookingservice/src/main/java/com/microservice/bookingservice/repository/ScheduleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.microservice.bookingservice.entities.Schedule;
 
 
-public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
+public interface ScheduleRepository extends JpaRepository<Schedule, Integer> {
 }
